[Refactor] Add `BaseModelConfig`
XTuner follows the principle that all modules are constructed from the
top-level model config, which inevitably leads to lower-level modules
depending on upper-level interfaces.

If the model config were declared in the model module, it would cause sub-modules like attention to import the
model, resulting in circular import issues. For this reason, we choose to declare the base class for ModelConfig in
the base config.